+++ mods/auxilias-crossbow/tools/sync-icons.ps1
@@ -49,9 +49,9 @@ foreach ($icon in $icons) {
             try {
                 $graphics.CompositingMode = [System.Drawing.Drawing2D.CompositingMode]::SourceCopy
                 $graphics.CompositingQuality = [System.Drawing.Drawing2D.CompositingQuality]::HighQuality
-                $graphics.InterpolationMode = [System.Drawing.Drawing2D.InterpolationMode]::HighQualityBicubic
-                $graphics.PixelOffsetMode = [System.Drawing.Drawing2D.PixelOffsetMode]::HighQuality
-                $graphics.SmoothingMode = [System.Drawing.Drawing2D.SmoothingMode]::HighQuality
+                $graphics.InterpolationMode = [System.Drawing.Drawing2D.InterpolationMode]::NearestNeighbor
+                $graphics.PixelOffsetMode = [System.Drawing.Drawing2D.PixelOffsetMode]::Half
+                $graphics.SmoothingMode = [System.Drawing.Drawing2D.SmoothingMode]::None
                 $graphics.Clear([System.Drawing.Color]::Transparent)
                 $graphics.DrawImage($source, 0, 0, $RuntimeSize, $RuntimeSize)
             }